split and move app src
