+++ src/UI.py
@@ -9,9 +9,9 @@ from transformers import pipeline # for captioning
 from xformers.ops import MemoryEfficientAttentionFlashAttentionOp
 import time # for sleep in FAKE_AI
 
-import config
-import utils
-import analytics
+import src.config as config
+import src.utils as utils
+import src.analytics as analytics
 
 # if active much more log output and ability to switch and select the used generation model 
 DEBUG = False
@@ -385,37 +385,37 @@ def create_gradio_interface():
         #end if dislaimer 
         return app
 
-if __name__ == "__main__":
-    app = None
-    try:
-        if config.read_configuration() == None: print("read configuration failed")
-        model = config.get_model()
-        if model.endswith("safetensors"):
-            utils.download_file_if_not_existing(url=config.get_model_url(), local_path=model)
+# if __name__ == "__main__":
+#     app = None
+#     try:
+#         if config.read_configuration() == None: print("read configuration failed")
+#         model = config.get_model()
+#         if model.endswith("safetensors"):
+#             utils.download_file_if_not_existing(url=config.get_model_url(), local_path=model)
 
-        print ("initializing ai models")
+#         print ("initializing ai models")
 
-        device = "cuda" if torch.cuda.is_available() else "cpu"
-        print (f"running on {device}")
-        if FAKE_AI:
-            print ("ai deactivated")
-        else:
-            IMAGE_TO_IMAGE_PIPELINE = load_model()
-            IMAGE_TO_TEXT_PIPELINE = load_captioner()
+#         device = "cuda" if torch.cuda.is_available() else "cpu"
+#         print (f"running on {device}")
+#         if FAKE_AI:
+#             print ("ai deactivated")
+#         else:
+#             IMAGE_TO_IMAGE_PIPELINE = load_model()
+#             IMAGE_TO_TEXT_PIPELINE = load_captioner()
 
-        if config.is_analytics_enabled():
-            analytics._DEBUG = DEBUG
-            analytics.start()
-        title = config.get_app_title()
-        print (f"starting Server ""{title}""")
-        app = create_gradio_interface()
-        app.launch(
-            share = config.is_gradio_shared(), 
-            max_file_size=8*gr.FileSize.MB
-            )
-        analytics.stop()
-    except Exception as e:
-        print (e)
-    finally:
-        if app != None: app.close()
-        print ("app closed")
+#         if config.is_analytics_enabled():
+#             analytics._DEBUG = DEBUG
+#             analytics.start()
+#         title = config.get_app_title()
+#         print (f"starting Server ""{title}""")
+#         app = create_gradio_interface()
+#         app.launch(
+#             share = config.is_gradio_shared(), 
+#             max_file_size=8*gr.FileSize.MB
+#             )
+#         analytics.stop()
+#     except Exception as e:
+#         print (e)
+#     finally:
+#         if app != None: app.close()
+#         print ("app closed")
